requirements.txt
colored==1.3.5
aiohttp==2.3.6
pwntools==2.2.1
